requirements.txt
# Real-Time Voice Assistant Dependencies

# Core Audio Processing
pyaudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1
webrtcvad==2.0.10
scipy==1.11.4
numpy==1.24.3

# Speech Recognition
openai-whisper==20231117
faster-whisper==0.10.0
speech-recognition==3.10.1

# Natural Language Processing
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
spacy==3.7.2

# Text-to-Speech
TTS==0.22.0
pyttsx3==2.90
piper-tts==1.2.0

# Web Framework & API
fastapi==0.108.0
uvicorn==0.25.0
websockets==12.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.9.1
asyncio==3.4.3
loguru==0.7.2

# Performance Optimization
numba==0.58.1
cython==3.0.7

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
